+++ leftwm-core/src/handlers/goto_tag_handler.rs
@@ -24,9 +24,10 @@ impl State {
         self.focus_manager.workspace_mut(&mut self.workspaces)?.tag = new_tag;
         self.focus_tag(&tag_num);
         self.update_static();
-        self.layout_manager
+        todo!()
+        /*self.layout_manager
             .update_layouts(&mut self.workspaces, self.tags.all_mut());
-        Some(true)
+        Some(true)*/
     }
 }
